add .env and /current route
