+++ app.js
@@ -9,10 +9,13 @@ const app = express();
 
 const formatsLogger = app.get("env") === "development" ? "dev" : "short";
 
+require("dotenv").config();
+require("./config/config-passport");
+
 app.use(logger(formatsLogger));
 app.use(cors());
 app.use(express.json());
-require("./config/config-passport");
+
 app.use("/api/contacts", contactsRouter);
 app.use("/api/users", usersRouter);
