+++ config/config-passport.js
@@ -1,7 +1,9 @@
 const passport = require("passport");
 const passportJWT = require("passport-jwt");
 const User = require("../models/schemas/user");
-const secret = require("../controller/users");
+require("dotenv").config();
+
+const secret = process.env.SECRET;
 
 const ExtractJWT = passportJWT.ExtractJwt;
 const Strategy = passportJWT.Strategy;
@@ -12,7 +14,7 @@ const params = {
 
 passport.use(
   new Strategy(params, function (payload, done) {
-    User.findOne({ _id: payload.id })
+    User.find({ _id: payload.id })
       .then(([user]) => {
         if (!user) {
           return done(new Error("User not found"));
@@ -22,22 +24,3 @@ passport.use(
       .catch((err) => done(err));
   })
 );
-
-const auth = (req, res, next) => {
-  passport.authenticate("jwt", { session: false }, (err, user) => {
-    if (!user || err) {
-      return res.status(401).json({
-        status: "error",
-        code: 401,
-        message: "Unauthorized",
-        data: "Unauthorized",
-      });
-    }
-    req.user = user;
-    next();
-  })(req, res, next);
-};
-
-module.exports = {
-  auth,
-};
