+++ package.json
@@ -12,6 +12,7 @@
     "bcryptjs": "^2.4.3",
     "cors": "2.8.5",
     "cross-env": "7.0.3",
+    "dotenv": "^16.3.1",
     "express": "4.17.1",
     "joi": "^17.9.2",
     "jsonwebtoken": "^9.0.2",
